001 (#12)
* Create Triangle_Angle_Type

* Add files via upload

* Create TruckTour.cpp

* Create RansomNote.cpp

Harold is a kidnapper who wrote a ransom note, but now he is worried it will be traced back to him through his handwriting. He found a magazine and wants to know if he can cut out whole words from it and use them to create an untraceable replica of his ransom note. The words in his note are case-sensitive and he must use only whole words available in the magazine. He cannot use substrings or concatenation to create the words he needs.

Given the words in the magazine and the words in the ransom note, print Yes if he can replicate his ransom note exactly using whole words from the magazine; otherwise, print No.

* Create SSSP.cpp

Complete the given code to implement the Bellman Ford Algorithm to find the Single-Source Shortest Path. Considering, node 0 as the source vertex, calculate shortest paths from it to all other vertices.

You have to define the BellmanFord() function. You may also implement additional utility functions as needed.

Use the value 100000 in place of inifinty (to represent disconnected vertices). Do not modify the main function.

* Create JourneyToTheMoon.py

The member states of the UN are planning to send 2 people to the moon. They want them to be from different countries. You will be given a list of pairs of astronaut ID's. Each pair is made of astronauts from the same country. Determine how many pairs of astronauts from different countries they can choose from.

* Create Prims(MST).cpp

Given a graph which consists of several edges connecting its nodes, find a subgraph of the given graph with the following properties:

The subgraph contains all the nodes present in the original graph.
The subgraph is of minimum overall weight (sum of all edges) among all such subgraphs.
It is also required that there is exactly one, exclusive path between any two nodes of the subgraph.
One specific node S  is fixed as the starting point of finding the subgraph using Prim's Algorithm.
Find the total weight or the sum of all edges in the subgraph.